move mvc config
