--- celements-servlet/src/main/java/com/celements/servlet/CelSpringWebConfig.java
@@ -1,11 +1,0 @@
-package com.celements.servlet;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-@EnableWebMvc
-public class CelSpringWebConfig implements WebMvcConfigurer {
-
-}
