+++ celements-xwiki-core/src/main/java/com/celements/spring/mvc/CelMvcInterceptor.java
@@ -1,0 +1,5 @@
+package com.celements.spring.mvc;
+
+import org.springframework.web.servlet.HandlerInterceptor;
+
+public interface CelMvcInterceptor extends HandlerInterceptor {}
